% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{combine_summary_info}
\alias{combine_summary_info}
\title{Combine Summary Information}
\usage{
combine_summary_info(summary_info)
}
\arguments{
\item{summary_info}{A list of summary information objects to be combined}
}
\value{
A combined summary information object
}
\description{
This function combines summary information from multiple studies for overlapping plots.
}
\examples{
\dontrun{
# Example usage
combined_summary <- combine_summary_info(summary_list)
}
}
