% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_nifti}
\alias{create_nifti}
\title{Create NIfTI File}
\usage{
create_nifti(nifti_template, data, mask)
}
\arguments{
\item{nifti_template}{A loaded NIfTI template file containing metadata}

\item{data}{A 1D vector of data values extracted from the mask}

\item{mask}{An nD mask array used to extract the original data}
}
\value{
A NIfTI object with data inserted back into the mask structure
}
\description{
This function creates a NIfTI file from a template, data, and mask.
}
\examples{
\dontrun{
# Example usage
nii <- create_nifti(template, effect_data, brain_mask)
}
}
