% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_plots}
\alias{create_plots}
\title{Create plots}
\usage{
create_plots(
  plot_data_list,
  plot_type = "simci",
  add_description = FALSE,
  save = FALSE,
  out_path = "output",
  file_name = "plot"
)
}
\arguments{
\item{plot_data_list}{A list of lists containing effect size data for plotting:
\itemize{
\item plot_data_list\link{i} = \code{plot_data}: contains:
\itemize{
\item \code{data}: A list containing sorted & downsampled effect size data & helpers: estimate, cons_estimate, lb, ub, below_cross_idx, and above_cross_idx
\item \code{extra_study_details}: A list containing extra descriptive info: percent_not_zero, max_cons_effect, group_by_title, n_title, mv_estimate, and mv_ci
\item \code{study_details}: A list of original study details: orig_stat_type, test_component_1, test_component_2, dataset, map_type, group, and ref
}
}}

\item{plot_type}{String; which plot type to make. Options are 'density' or 'simci'. Default is "simci".}

\item{add_description}{Logical; whether to add a description to the plot. Default is \code{FALSE}.}

\item{save}{Logical; whether to save the plot as a PNG file. Default is \code{FALSE}.}

\item{out_path}{A string specifying the output directory for saving plots. Default is "output".}

\item{file_name}{A string for the saved file name. Default is "plot".}
}
\value{
A plot visualizing effect sizes and simulated CIs.
}
\description{
This function is the master plotter for making effect size plots for individual
studies, overlapping studies, or meta-analytic results.
}
\examples{
# Example usage
# create_plots(pd)
}
