% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R, R/plot_helpers.R
\name{plot_connectivity_panel}
\alias{plot_connectivity_panel}
\title{Plot Connectivity Maps}
\usage{
plot_connectivity_panel(pp, plot_data_list, threshold_category = NA)

plot_connectivity_panel(pp, plot_data_list, threshold_category = NA)
}
\arguments{
\item{pp}{A list of plot parameters}

\item{plot_data_list}{A list containing effect size data for plotting:
\itemize{
\item \code{data}: A list containing sorted & downsampled effect size data & helpers: estimate, cons_estimate, lb, ub, below_cross_idx, and above_cross_idx
\item \code{summary_info}: A list containing extra descriptive info: percent_not_zero, max_cons_effect, group_by_title, and n_title
\item \code{mv_data}: A list containing multivariate effect size data: estimate, lb, and ub
\item \code{study_details}: A list of original study details: orig_stat_type, test_component_1, test_component_2, dataset, map_type, group, and ref
}}

\item{threshold_category}{Optional threshold category for data filtering (default: NA)}
}
\value{
A ggplot object containing connectivity matrix plots

A ggplot object containing one or overlapping densities
}
\description{
This function creates functional connectivity matrix plots.
}
\examples{
\dontrun{
# Example usage
plot_connectivity_panel(pp, plot_data)
}
Example usage
plot_connectivity_panel(plot_data)
}
