% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{plot_full_mat}
\alias{plot_full_mat}
\title{Plot Full Matrix}
\usage{
plot_full_mat(pp, triangle_ordered, ukb = FALSE, mapping_path = NA)
}
\arguments{
\item{pp}{A list of plot parameters}

\item{triangle_ordered}{A vector of triangle-ordered connectivity values}

\item{ukb}{Logical indicating if using UKB data (default: FALSE)}

\item{mapping_path}{Path to the mapping file for node labels (default: NA)}
}
\value{
A ggplot object containing the full matrix heatmap
}
\description{
This function creates a full connectivity matrix plot from triangle-ordered data.
}
\examples{
\dontrun{
# Example usage
plot_full_mat(pp, triangle_data, mapping_path = "path/to/mapping.csv")
}
}
