% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_single_plots.R
\name{plot_one_spatial}
\alias{plot_one_spatial}
\title{Single spatial effect size plot}
\usage{
plot_one_spatial(
  study_name,
  v_data,
  estimate = "d",
  motion_method = "none",
  pooling_method = "none",
  add_description = TRUE,
  do_minimal_title = FALSE
)
}
\arguments{
\item{study_name}{The name of the study to plot (string).}

\item{v_data}{A data list containing effect size data
(download data from OSF: https://doi.org/10.17605/OSF.IO/CWNJD, or format your data to the same format)}

\item{estimate}{The effect size measure to use ('d' or 'r_sq'). Default is 'd'.}

\item{motion_method}{The motion correction method used ('none', 'regression', 'threshold' for 0.1mm threshold,
or 'threshold2' for 0.2mm threshold). Default is 'none'.}

\item{pooling_method}{The pooling method used ('none', or 'net' for network-based pooling). Default is 'none'.}

\item{add_description}{Logical indicating whether to add a description to the plot. Default is TRUE.}

\item{do_minimal_title}{Logical indicating whether to use a minimal title. Default is FALSE.}
}
\value{
A ggplot object representing the spatial effect size plot.
}
\description{
This function plots a single spatial effect size plot.
}
\examples{
# Example usage
\dontrun{
plot_one_spatial("abcd_fc_r_rest_bmi", v_data = v) 
# v downloaded from OSF: https://doi.org/10.17605/OSF.IO/CWNJD
}

}
