% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_ci.R
\name{plot_sim_ci}
\alias{plot_sim_ci}
\title{Plot Simulated Confidence Intervals}
\usage{
plot_sim_ci(
  data,
  name,
  study_details,
  combo_name,
  mv_combo_name,
  group_by = "none",
  estimate = "d",
  save = FALSE,
  out_path = "output",
  file_name = "plot"
)
}
\arguments{
\item{data}{A list containing effect size data (e.g., \code{v$d_clean[[i]]}).}

\item{name}{A string specifying the study name (e.g., \code{names(v$d_clean[i])}).}

\item{study_details}{A list of study details (e.g., \code{v$study[i, ]}).}

\item{combo_name}{A string specifying the combo to plot.}

\item{mv_combo_name}{A string specifying the multivariate combo to plot.}

\item{group_by}{A string to specify grouping: "none", "orig_stat_type", or "category".}

\item{estimate}{A string to specify the effect size estimate: "d" or "r_sq"}

\item{save}{Logical; whether to save the plot as a PNG file. Default is \code{FALSE}.}

\item{out_path}{A string specifying the output directory for saving plots. Default is "output".}

\item{file_name}{A string for the saved file name. Default is "plot".}
}
\value{
A plot visualizing effect sizes and simulated CIs.
}
\description{
This function plots effect sizes (Cohen's d) and simulated confidence intervals (CIs)
for a given dataset. It allows optional grouping, visualization, and file saving.
}
\examples{
# Example usage
# plot_sim_ci(data = data, name = "abcd_fc_r_rest_bmi_z", study_details = study, 
#            combo_name = "pooling.none.motion.none.mv.none", mv_combo_name = "pooling.none.motion.none.mv.multi")
}
