% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_simci_panel}
\alias{plot_simci_panel}
\title{Function: Plot Simultaeous CIs}
\usage{
plot_simci_panel(pp, plot_data_list)
}
\arguments{
\item{pp}{A list of plot parameters}

\item{plot_data_list}{A list containing effect size data for plotting:
\itemize{
\item \code{data}: A list containing sorted & downsampled effect size data & helpers: estimate, cons_estimate, lb, ub, below_cross_idx, and above_cross_idx
\item \code{summary_info}: A list containing extra descriptive info: percent_not_zero, max_cons_effect, group_by_title, and n_title
\item \code{mv_data}: A list containing multivariate effect size data: estimate, lb, and ub
\item \code{study_details}: A list of original study details: orig_stat_type, test_component_1, test_component_2, dataset, map_type, group, and ref
}}
}
\value{
A ggplot object containing one or overlapping densities
}
\description{
Function: Plot Simultaeous CIs
}
\examples{
# Example usage
# plot_simci_panel(plot_data)
}
