% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_plot.R
\name{prep_data_for_plot}
\alias{prep_data_for_plot}
\title{Prepare Effect Size Data for Plot}
\usage{
prep_data_for_plot(
  data,
  study_details,
  combo_name,
  mv_combo_name,
  estimate = "d",
  plot_info = "NA",
  prep_spatial = FALSE,
  brain_masks = NA
)
}
\arguments{
\item{data}{A list containing effect size data (e.g., \code{v$d_clean[[i]]}).}

\item{study_details}{A list of study details (e.g., \code{v$study[i, ]}).}

\item{combo_name}{A string specifying the combo to plot.}

\item{mv_combo_name}{A string specifying the multivariate combo to plot.}

\item{estimate}{A string to specify the effect size estimate: "d" or "r_sq"}

\item{plot_info}{A list containing extra plot information (group_var, level, and reference atlas)}

\item{prep_spatial}{Logical; whether to prepare data for spatial plotting. Default is FALSE.}

\item{brain_masks}{A list containing brain mask information for spatial plotting}
}
\value{
A plot visualizing effect sizes and simulated CIs.
}
\description{
This function prepares a given study for effect size plots.
}
\examples{
# Example usage
\dontrun{
plot_info <- list(grouping_var = 'none', group_level = 'test', ref = 'atlas')
pd <- prep_data_for_plot(data = v$data[[i]], study_details = v$study[i, ], 
plot_info = plot_info, grouping_var='none', 
combo_name = "pooling.none.motion.none.mv.none", 
mv_combo_name = "pooling.none.motion.none.mv.multi")
}
}
