% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_spatial_plot.R
\name{prep_data_for_spatial_plot}
\alias{prep_data_for_spatial_plot}
\title{Prepare Effect Size Data for Plot}
\usage{
prep_data_for_spatial_plot(
  data,
  brain_masks,
  study_details,
  combo_name,
  mv_combo_name,
  estimate = "d",
  plot_info = "NA"
)
}
\arguments{
\item{data}{A list containing effect size data (e.g., \code{v$d_clean[[i]]}).}

\item{study_details}{A list of study details (e.g., \code{v$study[i, ]}).}

\item{combo_name}{A string specifying the combo to plot.}

\item{mv_combo_name}{A string specifying the multivariate combo to plot.}

\item{estimate}{A string to specify the effect size estimate: "d" or "r_sq"}

\item{plot_info}{A list containing extra plot information (group_var, level, and reference atlas)}
}
\value{
A plot visualizing effect sizes and simulated CIs.
}
\description{
This function prepares a given study for effect size plots.
}
\examples{
Example usage
pd <- prep_data_for_spatial_plot(data = v$data[[i]], study_details = v$study[i, ], plot_info$grouping_var='none',
             combo_name = "pooling.none.motion.none.mv.none", mv_combo_name = "pooling.none.motion.none.mv.multi")
}
