% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v}
\alias{v}
\title{Example BrainEffeX data to use with the vignettes}
\format{
A list with brain data, study information, and brain masks
}
\source{
This is a subset of the data available at \url{https://doi.org/10.17605/OSF.IO/CWNJD}
}
\usage{
v
}
\description{
Example BrainEffeX data to use with the vignettes
}
\author{
Hallee Shearer
}
\keyword{datasets}
