% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{add_plot_description}
\alias{add_plot_description}
\title{Add summary labels}
\usage{
add_plot_description(p, pp, summary_info, add_extra_text, do_minimal_title)
}
\arguments{
\item{p}{A ggplot object to which the labels will be added.}

\item{pp}{A list of plotting parameters.}

\item{summary_info}{A list of summary info for the plot.}

\item{add_extra_text}{Whether to add extra text below plot with more info.}

\item{do_minimal_title}{Whether to use minimal title option.}
}
\value{
A ggplot object with summary labels.
}
\description{
This function adds summary labels to a ggplot object.
}
\examples{
# Example usage
add_plot_description(p, pp, study_details, extra_study_details)
}
