% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot_description.R, R/plot_helpers.R
\name{add_plot_description}
\alias{add_plot_description}
\title{Add summary labels}
\usage{
add_plot_description(p, pp, summary_info, add_extra_text, do_minimal_title)

add_plot_description(p, pp, summary_info, add_extra_text, do_minimal_title)
}
\arguments{
\item{p}{A ggplot object to which the labels will be added.}

\item{study_details}{A list of original study details: orig_stat_type, test_component_1, test_component_2, dataset, map_type, group, and ref.}

\item{extra_study_details}{A list containing extra descriptive info: percent_not_zero, max_cons_effect, group_by_title, n_title, mv_estimate, and mv_ci.}
}
\value{
A ggplot object with summary labels.

A ggplot object with summary labels.
}
\description{
This function adds summary labels to a ggplot object.

This function adds summary labels to a ggplot object.
}
\examples{
# Example usage
# add_plot_description(p, pp, study_details, extra_study_details)
# Example usage
# add_plot_description(p, pp, study_details, extra_study_details)
}
