% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R, R/plot_helpers.R
\name{get_summary_info}
\alias{get_summary_info}
\title{Get Summary Information}
\usage{
get_summary_info(study_details, extra_study_details)

get_summary_info(study_details, extra_study_details)
}
\arguments{
\item{study_details}{A list of original study details: orig_stat_type, test_component_1, test_component_2, dataset, map_type, group, and ref.}

\item{extra_study_details}{A list containing extra descriptive info: percent_not_zero, max_cons_effect, group_by_title, n_title, mv_estimate, and mv_ci.}

\item{p}{A ggplot object to which the labels will be added.}
}
\value{
A list containing formatted summary information including title_text and bottom_text

A ggplot object with summary labels.
}
\description{
This function extracts and formats summary information from study details.

This function gets summary info to a ggplot object.
}
\examples{
\dontrun{
# Example usage
summary <- get_summary_info(study_details, extra_study_details)
}
Example usage
get_summary_info(pp, study_details, extra_study_details)
}
