% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_plot.R
\name{prep_data_for_plot}
\alias{prep_data_for_plot}
\title{Prepare Effect Size Data for Plot}
\usage{
prep_data_for_plot(
  data,
  name,
  study_details,
  combo_name,
  mv_combo_name,
  group_by_title = "None",
  estimate = "d"
)
}
\arguments{
\item{data}{A list containing effect size data (e.g., \code{v$d_clean[[i]]}).}

\item{name}{A string specifying the study name (e.g., \code{names(v$d_clean[i])}).}

\item{study_details}{A list of study details (e.g., \code{v$study[i, ]}).}

\item{combo_name}{A string specifying the combo to plot.}

\item{mv_combo_name}{A string specifying the multivariate combo to plot.}

\item{group_by_title}{A string to specify grouping: "none", "orig_stat_type", or "category".}

\item{estimate}{A string to specify the effect size estimate: "d" or "r_sq"}
}
\value{
A plot visualizing effect sizes and simulated CIs.
}
\description{
This function prepares a given study for effect size plots.
}
\examples{
# Example usage
# plot_sim_ci(data = data, name = "abcd_fc_r_rest_bmi_z", study_details = study,
#            combo_name = "pooling.none.motion.none.mv.none", mv_combo_name = "pooling.none.motion.none.mv.multi")
pd <- prep_data_for_plot(data = v$data[[i]], name = names(v$data[i]), study_details = v$study[i, ], group_by_title='none',
             combo_name = "pooling.none.motion.none.mv.none", mv_combo_name = "pooling.none.motion.none.mv.multi")
}
